+++ knp/backends/gpu/cuda-backend/impl/cuda_lib/register_all.cuh
@@ -11,7 +11,7 @@ REGISTER_CUDA_VECTOR_TYPE(knp::backends::gpu::cuda::CUDABackendImpl::ProjectionV
 REGISTER_CUDA_VECTOR_TYPE(knp::backends::gpu::cuda::CUDAPopulation<knp::neuron_traits::BLIFATNeuron>); \
 REGISTER_CUDA_VECTOR_TYPE(knp::backends::gpu::cuda::CUDAProjection<knp::synapse_traits::DeltaSynapse>); \
 REGISTER_CUDA_VECTOR_TYPE(knp::backends::gpu::cuda::CUDAProjection<knp::synapse_traits::DeltaSynapse>::Synapse); \
-REGISTER_CUDA_VECTOR_TYPE(uint64_t); \
+REGISTER_CUDA_VECTOR_TYPE(unsigned long long); \
 REGISTER_CUDA_VECTOR_TYPE(knp::backends::gpu::cuda::Subscription); \
 REGISTER_CUDA_VECTOR_TYPE(knp::backends::gpu::cuda::MessageVariant); \
 REGISTER_CUDA_VECTOR_TYPE(knp::backends::gpu::cuda::UID)
